requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
click==8.1.7
dnspython==2.5.0
fastapi==0.109.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pymongo==4.6.1
python-dotenv==1.0.1
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0
